Source_Code/requirements.txt
Pillow==10.1.0
pyinstaller==6.3.0
